Refactor for better readability
