+++ src/main/kotlin/com/block/gooseintellij/actions/AskGooseAboutFileOrDirectoryAction.kt
@@ -1,56 +1,16 @@
-package com.block.gooseintellij
+package com.block.gooseintellij.actions
 
 import com.block.gooseintellij.toolWindow.GooseTerminalWidget
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
-import com.intellij.openapi.ui.Messages
-import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.progress.Task
-import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.ui.Messages
+import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.openapi.diagnostic.Logger
-
-class OpenGooseTerminalAction : AnAction() {
-    override fun actionPerformed(event: AnActionEvent) {
-        val project = event.project
-        val toolWindow = ToolWindowManager.getInstance(project!!).getToolWindow("Goose Terminal")
-
-        if (toolWindow != null && !toolWindow.isVisible) {
-            toolWindow.activate(null)
-            toolWindow.show()
-        }
-    }
-}
-
-const val GOOSE_COMMAND_FORMAT = "Please analyze the following code %s\\n in file: %s\\n"
-
-class SendToGooseAction : AnAction() {
-    override fun actionPerformed(event: AnActionEvent) {
-        val project = event.project
-        val toolWindow = ToolWindowManager.getInstance(project!!).getToolWindow("Goose Terminal")
-
-        if (toolWindow == null || !toolWindow.isVisible) {
-            OpenGooseTerminalAction().actionPerformed(event)
-        }
-
-        val contentManager = toolWindow?.contentManager
-        val content = contentManager?.getContent(0)
-        val gooseTerminal = content?.component as? GooseTerminalWidget
-
-        val editor = event.getData(CommonDataKeys.EDITOR)
-        val selectedText = editor?.selectionModel?.selectedText
-        val psiFile = event.getData(CommonDataKeys.PSI_FILE)
-
-        if (selectedText != null && selectedText.isNotEmpty() && psiFile != null) {
-            val filePath = psiFile.virtualFile.path
-            val command = String.format(GOOSE_COMMAND_FORMAT, selectedText.replace("\n", "\\n"), filePath)
-            GooseTerminalWidget.writeCommandToTerminal(gooseTerminal?.connector!!, command)
-        } else {
-            Messages.showMessageDialog("No file or text selected.", "Warning", Messages.getWarningIcon())
-        }
-    }
-}
+import com.block.gooseintellij.utils.GooseUtils
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 
 class AskGooseAboutFileOrDirectoryAction : AnAction() {
 
@@ -60,6 +20,13 @@ class AskGooseAboutFileOrDirectoryAction : AnAction() {
         logger.info("ActionPerformed triggered")
         
         val project = event.project
+        
+        // Check Goose availability
+        if (GooseUtils.getGooseState() != true && GooseUtils.getSqGooseState() != true) {
+            Messages.showMessageDialog(project, "Goose or SQ Goose is not available. Please install them.", "Error", Messages.getErrorIcon())
+            return
+        }
+
         val virtualFile = event.getData(CommonDataKeys.VIRTUAL_FILE)
         logger.info("Project: $project, VirtualFile: $virtualFile")
 
@@ -84,7 +51,7 @@ class AskGooseAboutFileOrDirectoryAction : AnAction() {
                             val gooseTerminal = content?.component as? GooseTerminalWidget
 
                             if (gooseTerminal != null) {
-                                GooseTerminalWidget.writeCommandToTerminal(gooseTerminal.connector!!, question)
+                                GooseUtils.writeCommandToTerminal(gooseTerminal.connector!!, question)
                                 logger.info("Question sent to GooseTerminalPanel")
                             } else {
                                 logger.warn("GooseTerminalPanel is null")
